+++ esp-hal/src/assist_debug.rs
@@ -52,8 +52,7 @@ impl<'d> DebugAssist<'d> {
         for core in crate::system::Cpu::other() {
             crate::interrupt::disable(core, Interrupt::ASSIST_DEBUG);
         }
-        unsafe { crate::interrupt::bind_interrupt(Interrupt::ASSIST_DEBUG, handler.handler()) };
-        crate::interrupt::enable(Interrupt::ASSIST_DEBUG, handler.priority());
+        crate::interrupt::bind_handler(Interrupt::ASSIST_DEBUG, handler);
     }
 
     fn regs(&self) -> &pac::assist_debug::RegisterBlock {